backend/requirements.txt
# Django 6.x: Python 3.12+; при MariaDB нужна 10.6+ (не 10.5).
Django>=6.0,<7
djangorestframework>=3.17,<4
django-cors-headers>=4.9,<5
djangorestframework-simplejwt>=5.5,<6
django-filter>=25.2,<26
django-unfold>=0.87,<1
django-simple-captcha>=0.6.3,<0.7
pytest>=9,<10
pytest-django>=4.12,<5
Pillow>=11,<12
